requirements/base.txt
Eve==1.1.1
environs==8.0.0
py-bcrypt==0.4
